asset-build-refactor: refactor: move build and tests script to larva
