+++ package.json
@@ -3,13 +3,17 @@
   "private": true,
   "scripts": {
 	"test": "./node_modules/.bin/jest",
-	"larva": "nodemon -e twig,js,html packages/larva/lib/server.js"
+	"larva": "nodemon -e twig,js,html packages/larva/lib/server.js",
+	"watch": "node packages/larva/lib/watch.js"
   },
   "devDependencies": {
     "express": "^4.17.1",
     "jest": "^24.8.0",
     "lerna": "^3.15.0",
 	"twing": "^2.2.7",
-	"stylelint": "^10.1.0"
+	"stylelint": "^10.1.0",
+	"gaze": "^1.1.3",
+    "globby": "^10.0.0",
+    "node-sass": "^4.12.0"
   }
 }
