--- packages/css-utilities/lib/utils/utils.js
@@ -1,6 +1,0 @@
-
-module.exports = {
-	getScssPathsWithExtension: require( './getScssPathsWithExtension' ),
-	concatenateFileData: require( './concatenateFileData' ),
-	renderSass: require( './renderSass' ),
-}
