+++ packages/css-utilities/package.json
@@ -20,8 +20,7 @@
     "url": "git+https://github.com/penske-media-corp/pmc-larva.git"
   },
   "scripts": {
-    "test": "../../node_modules/.bin/jest",
-    "watch": "node index.js"
+    "test": "../../node_modules/.bin/jest"
   },
   "bugs": {
     "url": "https://github.com/penske-media-corp/pmc-larva/issues"
@@ -31,10 +30,6 @@
   },
   "devDependencies": {
     "@penskemediacorp/larva": "^1.3.0-alpha.0",
-    "@penskemediacorp/larva-scss": "^1.3.0-alpha.0",
-    "gaze": "^1.1.3",
-    "globby": "^10.0.0",
-    "node-sass": "^4.12.0",
-    "stylelint": "^10.1.0"
+    "@penskemediacorp/larva-scss": "^1.3.0-alpha.0"
   }
 }
