+++ packages/larva/__tests__/lib/utils/concatenateFileData.test.js
@@ -7,8 +7,8 @@ const fixture = path.resolve( __dirname, '../../fixtures' );
 describe( 'concatenateFileData', () => {
 	it( 'concats data from files in array', () => {
 		let paths = [
-			path.join( fixture, './src/u-test.common.inline.scss' ),
-			path.join( fixture, './src/u-test-two.common.inline.scss' ),
+			path.join( fixture, './src/patterns/07-utilities/u-test.common.inline.scss' ),
+			path.join( fixture, './src/patterns/07-utilities/u-test-two.common.inline.scss' ),
 		];
 
 		let stringsExpected = [
