+++ packages/larva/__tests__/lib/utils/getScssPathsWithExtension.test.js
