+++ packages/larva/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = {
+	build: require( './lib/build' ),
+	watch: require( './lib/watch' )
+};
