--- packages/larva/lib/larva.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = larva;
-
-function larva() {
-	// TODO
-}
