+++ packages/larva/lib/utils/getScssPathsWithExtension.js
