+++ packages/larva/package.json
@@ -5,7 +5,7 @@
   "author": "laras126 <laras126@gmail.com>",
   "homepage": "https://github.com/Penske-Media-Corp/pmc-larva/tree/master/packages/larva#readme",
   "license": "MIT",
-  "main": "lib/larva.js",
+  "main": "index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Penske-Media-Corp/pmc-larva.git"
